Merge V1 into master (#83)
* Minor initial styling

* Support specific base map for AK Fires

This overrides the base layer for 3338, and adds the concept of per-map base map configuration options (restricting bounds, zoom, and other base Leaflet options).

* Adds scaffolding to show specific layers for a map by default

Again, this points the way to a general refactoring for per-map specific configuration; this is an interim measure for V1!

* Adds per-map class to body element

The name of the map is slugified and added as a class to the body element so that page-level themes can be overridden per-map.  More a proof of concept that will need proper encapsulation later!

* Minor initial styling

* Adds per-map class to body element

The name of the map is slugified and added as a class to the body element so that page-level themes can be overridden per-map.  More a proof of concept that will need proper encapsulation later!

* Splash screen now shows map abstract & continue button

* Modified the expected layer that should be enabled by default for the Active Fires map.