Assign max zoom level, fix zoom jump bug
When the map was zoomed in, it would 'skip' to an invalid zoom level.  This was because the `resolutions` array defined in the CRS didn't have enough elements in it and so things sort of broke down in Proj4Leaflet.  This isn't exactly a bug in that code, though there could be more meaningful errors thrown, because there is a definite and purposeful relationship between the `resolutions` array and the zoom levels.  I adjusted the zoom levels and expanded the resolutions array to address the bug.