Implement basic WFS query to render GeoJSON layer
Two stubs are added to the code, one is called and performs a WFS query to the server to fetch all current fire polygons, which are then added to the map with clickable popups.

Also in this commit is a change that should prevent Bower from failing when switching branches (in the commit hook), added to this branch so I could rebase the commits correctly.