Merge V2 (#92)
* Assign max zoom level, fix zoom jump bug (#84)

* Assign max zoom level, fix zoom jump bug

When the map was zoomed in, it would 'skip' to an invalid zoom level.  This was because the `resolutions` array defined in the CRS didn't have enough elements in it and so things sort of broke down in Proj4Leaflet.  This isn't exactly a bug in that code, though there could be more meaningful errors thrown, because there is a definite and purposeful relationship between the `resolutions` array and the zoom levels.  I adjusted the zoom levels and expanded the resolutions array to address the bug.

* Change to name of active_fires layer in basemap to allow for services to load properly.

* relax bounds to show more of Canada (#85)

* Removed distribution_url metadata from sidebar (#86)

* Implement basic WFS query to render GeoJSON layer

Two stubs are added to the code, one is called and performs a WFS query to the server to fetch all current fire polygons, which are then added to the map with clickable popups.

Also in this commit is a change that should prevent Bower from failing when switching branches (in the commit hook), added to this branch so I could rebase the commits correctly.

* Uses the configuration environment variables to set the Geoserver URL for the WFS URL for the active fires layer.

* Add markers to fire polygons (#90)

A few other minor changes:

 * Modify style so legend will always fit inside info sidebar
 * Fix possible issue with the `grunt-contrib-imagemin` package breaking the build

* Rework template (#91)

* Rework template used to display the map

 * Remove unused top page navigation
 * Relax formatting inside the layers tool
 * Title of page is title of map
 * Remove borders when in dual map sync mode

* Fix position of footer

* pin map height

* Modified the amount of longitude available to the basemap.