Split global v. per-map features into different controllers
This moves the Fire-map-specific code into its own controller and keeps common code in the top-level MapCtrl object.  Further refinement is necessary.

The `basemap` Service is commented out since its functionality is moved into the Fire-map-specific controller, but there are some useful notes in there for the pan-Arctic projection, which I'll properly remove/document in a future commit.