Add plugin "registry" service for customized map behavior
This completes the refactoring to allow new maps to have custom behavior, isolated in their own controllers.  Behavior for a default map is sketched into place but doesn't work yet (will explore for a future map), but it preserves some of the technical details needed to get the polar stereographic projections working.