Adds some easing to zoom when clicking on marker (#135)
Also, reduces the maximum zoom level so it's not fully zoomed in.  This is only a partial solution -- it does slow things down a bit but it's slightly clunky.  There really isn't a better overall option unless we switch to V1.x of Leaflet which isn't released yet and may not be fully compatible with the application, but it's worth doing some more study of that in the future.