Merge V3 (#138)
* Adds jscs, start conforming to style guide

Using [Google style guide](https://google.github.io/styleguide/javascriptguide.xml), because we're using Angular.  Adds [jscs](http://jscs.info) to the build process.  This commit also contains fixes that jscs was able to do automatically.

Adds a new grunt task, `grunt tidy` that runs jscs and can be adapted to run other linters in the future.

* Fixes style violations

* Adds jshint to `grunt tidy` task, fixes

Addresses warnings caught by `grunt tidy` with jshint added to that task.

* Update .jscsrc

remove contradictory maximum line length limit & ignore line length limit :)

* remove refs to "about" route, template

* Move all current map functionality to a sub-controller

Bulk copy/paste of all existing functionality moved to a new sub-Controller to sketch how out how make these dynamic.

Also removes cruft leftover with the About page.

* Split global v. per-map features into different controllers

This moves the Fire-map-specific code into its own controller and keeps common code in the top-level MapCtrl object.  Further refinement is necessary.

The `basemap` Service is commented out since its functionality is moved into the Fire-map-specific controller, but there are some useful notes in there for the pan-Arctic projection, which I'll properly remove/document in a future commit.

* Add plugin "registry" service for customized map behavior

This completes the refactoring to allow new maps to have custom behavior, isolated in their own controllers.  Behavior for a default map is sketched into place but doesn't work yet (will explore for a future map), but it preserves some of the technical details needed to get the polar stereographic projections working.

* Refactor tour to per-map implementation

The Tour directive now invokes a Service provided through the MapRegistry.  The per-map tour service, located in `/app/scripts/map/(map name)/tour.js` is a wrapper around the Tour itself.  The directive runs the tour and manages state.

* Adds remaining changes supporting this functionality

* Modify styling of splash page

* Scroll wheel allowed by default in the map configuration.

* Changed base map configuration function to point at the WMS service of GeoServer for the default and alaska-wildfires controllers.

* Created favicon and added it to MapVenture. This could (and likely should) be replaced by someone more artistic than myself. (#123)

* Modified the glyphicons from eyes to checkboxes. (#121)

* Slight changes to default map

The GeoServer instance needs to have the natural earth layer named ne_10m_coastline present.  This map displays in polar stereographic (EPSG:3572) just to preserve that code for the moment, may not be a good choice for default in the future.

* Fix tour, improve grunt watch process

The Tour has NCEP references stripped and works for the default map.  The Gruntfile was modified to also watch the `app/scripts/maps` directory for changes so that it will auto-reload when editing those files.

* add maps to proper build block in index.html (#132)

* Bug fix for minimize layer menu. (#134)

* Adds some easing to zoom when clicking on marker (#135)

Also, reduces the maximum zoom level so it's not fully zoomed in.  This is only a partial solution -- it does slow things down a bit but it's slightly clunky.  There really isn't a better overall option unless we switch to V1.x of Leaflet which isn't released yet and may not be fully compatible with the application, but it's worth doing some more study of that in the future.