Fixes #267, edge of map visible on default viewport for large screens.
The underlying issue here is that white border is actually the extent of the map, which is bounded according to (more or less) the extent of AK Albers on the server side.  You can adjust the Leaflet object's maxBounds, which will shuffle the default center of the map somewhat, but it's a little odd/fiddly.