Fixes #330, fire graph shows up twice.
I made a mistake when resolving a merge conflict.  This fixes that error and still fixes the bug with the close button in Firefox.  To test this, check the functionality of the graph in Firefox -- it should only show up once, be correctly formatted and the close button should work.